requirements.txt
streamlit>=1.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0